Capabilities Reorganization
Move the client and server capabilities out into dedicated files, as things were getting a little out of hand. Also fills in a few more holes in those structures.
